% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{tw_user_timeline}
\alias{tw_user_timeline}
\title{function for getting the twitter user timeline}
\usage{
tw_user_timeline(screen_name, n = 1, key = "", secret = "")
}
\description{
function for getting the twitter user timeline
}

