% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{tw_usr_tmln}
\alias{tw_usr_tmln}
\title{function for getting the twitter user timeline}
\usage{
tw_usr_tmln(screen_name, max_id = "", count = 200, key, secret)
}
\description{
function for getting the twitter user timeline
}

