% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{twl_set_credentials}
\alias{twl_set_credentials}
\title{function storing key and secret for an session}
\usage{
twl_set_credentials(key = "", secret = "", token = "",
  token_secret = "")
}
\arguments{
\item{key}{key to be used for authentication}

\item{secret}{secret to be used for authentication}
}
\description{
function storing key and secret for an session
}
\seealso{
\link{twl_reset_credentials}
}

