% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authentication.R
\docType{data}
\name{twl_storage}
\alias{twl_storage}
\title{Package storage for session specific information like credentials}
\format{\preformatted{<environment: 0x295aa70> 
}}
\usage{
twl_storage
}
\description{
Package storage for session specific information like credentials
}
\seealso{
\link{twl_token}, \link{twl_set_credetnials}, \link{twl_reset_credetnials}
}
\keyword{datasets}

