% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authentication.R
\name{twl_token}
\alias{twl_token}
\title{function doing twitter oauth authentication}
\usage{
twl_token(key = "", secret = "")
}
\arguments{
\item{key}{optional key argument; if left as is twitter_token() will first
look if twitterlight_key was set before in the session by
set_twitter_credetnials and thereafter check if it can find the values in
an .Renviron file via sys.getenv("twitterlight_key")}

\item{secret}{same as key parameter except function will search for
twitterlight_secret}
}
\value{
twitter token used for API interaction by package functions

  internal usage: \code{httr::GET(... httr::config( token = twitter_token(key, secret) ) )}
}
\description{
function doing twitter oauth authentication
}
\seealso{
\link{twl_set_credentials}, \link{Sys.getenv}, \link{oauth_app}, \link{oauth1.0_token}
}

