% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authentication.R
\docType{data}
\name{twl_storage}
\alias{twl_storage}
\title{Package storage for session specific information like credentials}
\format{\preformatted{<environment: 0x4471970> 
}}
\usage{
twl_storage
}
\description{
Package storage for session specific information like credentials
}
\keyword{datasets}

